+++ dbfiles/fall2023.json
@@ -1,0 +1,21 @@
+{
+  "tracks":[
+    "laguna seca",
+    "cota",
+    "spa",
+    "daytona oval",
+    "sonoma raceway"
+  ],
+  "cars":[
+    "corvette c7r",
+    "2013 v8 forumula 1",
+    "mazda 787b",
+    "vrc forumula na 1999",
+    "shelby cobra",
+    "nissan gt-r gt3",
+    "1982 ligier js8",
+    "mercedes clr",
+    "vrc indycar",
+    "next gen nascar"
+  ]
+}
