+++ src/main/java/com/github/loafabreadly/Main.java
@@ -5,6 +5,7 @@ import com.github.loafabreadly.command.Command;
 import com.github.loafabreadly.command.CreateSeriesCmd;
 import com.github.loafabreadly.command.ModifyCmd;
 import com.github.loafabreadly.command.RandomPickCmd;
+import com.github.loafabreadly.utils.Constants;
 import me.koply.kcommando.KCommando;
 import me.koply.kcommando.integration.impl.javacord.JavacordIntegration;
 import org.apache.logging.log4j.LogManager;
@@ -50,7 +51,6 @@ public class Main {
             kc.registerObject(new ModifyCmd());
             kc.registerObject(new RandomPickCmd());
             logger.info("Successfully registered our KC commands!");
-
             kc.build();
         }
         catch (Exception e) {
@@ -60,10 +60,9 @@ public class Main {
 
     //TODO Add in View command
     //TODO Add in series tracker JSON file
-    //TODO Add in randomizer command
 
     private static String getToken() {
-        String token = System.getenv("DISCORD_TOKEN");
+        String token = System.getenv("DISCORD_TOKEN").;
         logger.info("Env var was defined for token!");
         return Objects.requireNonNull(token);
     }
