+++ src/main/java/com/github/loafabreadly/command/CreateSeriesCmd.java
@@ -1,9 +1,9 @@
 package com.github.loafabreadly.command;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.loafabreadly.ErrorHandler;
+import com.github.loafabreadly.utils.ErrorHandler;
 import com.github.loafabreadly.Main;
-import com.github.loafabreadly.SeriesObject;
+import com.github.loafabreadly.structures.SeriesObject;
 import lombok.NonNull;
 import me.koply.kcommando.internal.OptionType;
 import me.koply.kcommando.internal.annotations.HandleSlash;
@@ -16,7 +16,6 @@ import org.javacord.api.interaction.callback.InteractionFollowupMessageBuilder;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.List;
 
 public class CreateSeriesCmd implements Command {
