+++ src/main/java/com/github/loafabreadly/command/ModifyCmd.java
@@ -1,10 +1,10 @@
 package com.github.loafabreadly.command;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.loafabreadly.ErrorHandler;
+import com.github.loafabreadly.utils.ErrorHandler;
 import com.github.loafabreadly.Main;
-import com.github.loafabreadly.SeriesObject;
-import com.github.loafabreadly.SimBoizEmbedBuilder;
+import com.github.loafabreadly.structures.SeriesObject;
+import com.github.loafabreadly.utils.SimBoizEmbedBuilder;
 import lombok.NonNull;
 import me.koply.kcommando.internal.OptionType;
 import me.koply.kcommando.internal.annotations.HandleSlash;
@@ -16,7 +16,6 @@ import org.javacord.api.interaction.SlashCommandInteraction;
 import org.javacord.api.interaction.callback.InteractionFollowupMessageBuilder;
 
 import java.io.File;
-import java.util.Iterator;
 import java.util.List;
 
 public class ModifyCmd implements Command {
