+++ src/main/java/com/github/loafabreadly/utils/BotFile.java
@@ -1,0 +1,9 @@
+package com.github.loafabreadly.utils;
+
+import java.io.File;
+
+public class BotFile extends File {
+    public BotFile(String pathname) {
+        super("dbfiles/" + pathname);
+    }
+}
